requirements.txt
numpy>=1.23.5
torch>=1.9.0
transformers>=4.34.1
miditoolkit>=1.0.0
tensorboard>=2.14.0
pyyaml>=6.0.1
tqdm>=4.66.1
pretty_midi